requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
pypdf
python-dotenv
sentence-transformers
huggingface-hub
